### 2020-10-23 #2771 @whedon
[PRE REVIEW]: gospl: Global Scalable Paleo Landscape Evolution
**Submitting author:** @tristan-salles (<a href="http://orcid.org/0000-0001-6095-7689">Tristan Salles</a>)
**Repository:** <a href="https://github.com/Geodels/gospl/" target ="_blank">https://github.com/Geodels/gospl/</a>
**Version:** v0.2.0
