### 2020-10-27 60e82d8
Update installDocker.rst
