+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy==1.19.2
+numpy==1.19.3
 scipy>=1.5.2
 Cython==0.29.21
 mpi4py==3.0.3
