+++ requirements.txt
@@ -9,7 +9,7 @@ ruamel.yaml==0.16.12
 fastfunc==0.2.3
 meshio>=4.2.2
 meshplex>=0.13.3
-pre-commit==2.7.1
+pre-commit==2.8.1
 vtk>=8.1.2
 numpy-indexed==0.3.5
 scikit-fuzzy==0.4.2
