### 2020-10-30 #2804 @kbarnhart
@tristan-salles could you add the DOI's identified as missing by whedon? There are also some parts of the bibliography (mostly titles) that could benefit from extra curly braces to enforce capitalization (useful reference [here](https://www.ece.ucdavis.edu/~jowens/biberrors.html)).
