### 2020-10-30 b952fa0
update DOIs for JOSS submission
