+++ paper/paper.bib
@@ -10,7 +10,7 @@
 
 @article{Helland:16,
     author = {Helland-Hansen, W. and Somme, T. O. and Martinsen, O. J. and Lunt, I. and Thurmond, J.},
-    title = "{Deciphering Earth's Natural Hourglasses: Perspectives On Source-To-Sink Analysis}",
+    title = "{D}eciphering {E}arth's {N}atural {H}ourglasses: {P}erspectives {O}n {S}ource-{T}o-{S}ink {A}nalysis",
     journal = {Journal of Sedimentary Research},
     volume = {86},
     number = {9},
@@ -21,7 +21,7 @@
 
 @article{Richardson:14,
   author = {Richardson, A. and Hill, C. N. and Perron, J. T.},
-  title = {IDA: An implicit, parallelizable method for calculating drainage area},
+  title = "{IDA}: {A}n implicit, parallelizable method for calculating drainage area",
   journal = {Water Resources Research},
   volume = {50},
   number = {5},
@@ -32,7 +32,7 @@
 
 @article{Barnes:14,
    author = {Barnes, R. and Lehman, C. and Mulla, D.},
-    title = {Priority-flood: An optimal depression-filling and watershed-labeling algorithm for digital elevation models},
+    title = "Priority-flood: {A}n optimal depression-filling and watershed-labeling algorithm for digital elevation models",
   journal = {Computers \& Geosciences},
      year = {2014},
    volume = {62},
@@ -42,7 +42,7 @@
 
 @article{Yuan:19,
   author = {Yuan, X.P. and Braun, J. and Guerit, L. and Simon, B. and Bovy, B. and Rouby, D. and Robin, C. and Jiao, R.},
-  title = {Linking continental erosion to marine sediment transport and deposition: A new implicit and O(N) method for inverse analysis},
+  title = "Linking continental erosion to marine sediment transport and deposition: {A} new implicit and {O(N)} method for inverse analysis",
   journal = {Earth and Planetary Science Letters},
   volume = {524},
   pages = {115728},
@@ -53,7 +53,7 @@
 
 @article{Salles:19,
   author = {Salles, T.},
-  title = {eSCAPE: Regional to Global Scale Landscape Evolution Model v2.0},
+  title = "e{SCAPE}: {R}egional to {G}lobal {S}cale {L}andscape {E}volution {M}odel v2.0",
   journal = {Geoscientific Model Development},
   volume = {12},
   year = {2019},
@@ -65,43 +65,47 @@
 
 @article{Braun:97,
   author = {Braun, J. and Sambridge, M.},
-  title = {Modelling landscape evolution on geological time scales: a new method based on irregular spatial discretization},
+  title = "Modelling landscape evolution on geological time scales: {A} new method based on irregular spatial discretization",
   journal = {Basin Research},
   volume = {9},
   number = {1},
   pages = {27--52},
-  year = {1997}
+  year = {1997},
+  doi = {10.1046/j.1365-2117.1997.00030.x}
 }
 
 @Inbook{Tucker:01,
   author = {Tucker, Gregory and Lancaster, Stephen  and Gasparini, Nicole and Bras, Rafael},
   editor = {Harmon, Russell S.  and Doe, William W.},
-  title = {The Channel-Hillslope Integrated Landscape Development Model (CHILD)},
+  title = "{T}he {C}hannel-{H}illslope {I}ntegrated {L}andscape {D}evelopment {M}odel ({CHILD})",
   bookTitle = {Landscape Erosion and Evolution Modeling},
   publisher = {Springer US},
   address = {Boston, MA},
   pages = {349--388},
-  year = {2001}
[…]
