+++ requirements.txt
@@ -3,7 +3,7 @@ scipy>=1.5.2
 Cython==0.29.21
 mpi4py==3.0.3
 petsc4py>=3.13.0
-h5py==2.10.0
+h5py>=2.10.0
 pandas>=1.1.2
 ruamel.yaml==0.16.12
 fastfunc==0.2.3
