### 2020-10-30 #2804 @tristan-salles
@kbarnhart I have updated the DOI's and the Titles capitalisation in the bibliography.
Thank you to the reviewers
