+++ paper/paper.bib
@@ -115,7 +115,8 @@
   volume = {35},
   number = {1},
   pages = {28--50},
-  year = {2010}
+  year = {2010},
+  doi = {10.1002/esp.1952}
 }
 
 @book{Granjeon:99,
