+++ gospl/mesher/unstructuredmesh.py
@@ -140,8 +140,12 @@ class UnstMesh(object):
         # Voronoi and simplices declaration
         Tmesh.create_edges()
         cc = Tmesh.cell_circumcenters
-        edges_nodes = Tmesh.edges["nodes"]
-        cells_nodes = Tmesh.cells["nodes"]
+        if meshplex.__version__ >= "0.1.14":
+            edges_nodes = Tmesh.edges["points"]
+            cells_nodes = Tmesh.cells["points"]
+        else:
+            edges_nodes = Tmesh.edges["nodes"]
+            cells_nodes = Tmesh.cells["nodes"]
         cells_edges = Tmesh.cells["edges"]
 
         # Finite volume discretisation
