### 2020-11-23 #2804 @kbarnhart
@johnjarmitage, @cmshobe I wanted to write with a quick check-in to see if you had any questions for me about the JOSS review process.
