+++ gospl/mesher/unstructuredmesh.py
@@ -140,7 +140,7 @@ class UnstMesh(object):
         # Voronoi and simplices declaration
         Tmesh.create_edges()
         cc = Tmesh.cell_circumcenters
-        if meshplex.__version__ >= "0.1.14":
+        if meshplex.__version__ >= "0.14.0":
             edges_nodes = Tmesh.edges["points"]
             cells_nodes = Tmesh.cells["points"]
         else:
