### 2020-11-24 #2804 @johnjarmitage
@kbarnhart yes:
1. What level of explanation of the equations/methods being solved within the software package is expected? I have tested one aspect of the code and it works as it should. I am however not sure exactly what is going on with the lateral movement of the tectonic plates. In a science journal this might be a problem, but from the check list above I don't see this being flagged as an issue.
2. (Related) If I have suggestions for improving the description, where do I put them? An issue in the code repository, or is there a formal review in JOSS?
