### 2020-11-25 #2804 @kbarnhart
@johnjarmitage great questions.

_Regarding the first_: The two most relevant elements of the review criteria here would be:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I would expect that the functional form of the equations being solved by the package are stated, their constants listed, a description of any boundary conditions, and a statement of numerical methods used, if any (e.g., solved with finite difference). I would expect that, if known solutions to the governing equations exist that some sort of test would exist to demonstrate the equations are being solved correctly. I would not expect that every equation be written out in a discretized form.

This information could be located anywhere in the package documentation (e.g., narrative documentation, examples, api).

If I understand your question correctly, I think you are indicating that may not be sufficient detail in describing how the tectonic boundary conditions are implemented on the discretized grid (e.g., warping of existing grid cells due to tectonic advection, creating or removing grid cells if they get too big or too small, what quantities are conserved). I would expect some description of how this is implemented and if there are any inputs exposed to the user that control this behavior.

Let me know if this answers your question.

_Regarding the second_: We recommend that you make most recommendations/requests for change as issues in the source repository and link them here (just paste "openjournals/joss-reviews/issues/2804" into the new issue and github will do the linking). Thus this issue compiles the major topics of discussion, requested changes that are motivated by this review, but does not contain them.
