### 2020-11-25 Geodels/gospl#94 @johnjarmitage
technical guide : description of plate tectonics?
As part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/2804), I am thinking that a description of how plate tectonics is included might be a great addition to the page https://gospl.readthedocs.io/en/latest/tech_guide/index.html#tech-guide. The erosion, deposition, and compaction algorithms are described, but I don't see a description of the methods used to:
* laterally move the surfaces (gplates?)
* remesh if needed, as the grids get distorted
* deal with the collision of sediments (I assume there is no actual compression of sediment and crust, but something must be done at the plate boundaries, right?)

Or if all of the above is wrong, then maybe some text explaining what is done, because it is possible I am assuming a complexity that does not exist.
