### 2020-11-25 #2804 @kbarnhart
@johnjarmitage looks great to me. Now @tristan-salles can respond to your common on the new issue Geodels/gospl#94 and if discussion is needed it can happen there.
