### 2020-11-26 61e1727
add description of plate motion
