### 2020-11-26 Geodels/gospl#94 @tristan-salles
Following @johnjarmitage suggestion I have added a description of how tectonic displacements are accounted for in the code: https://gospl.readthedocs.io/en/latest/tech_guide/tecto.html
