### 2020-11-30 Geodels/gospl#94 @johnjarmitage
Thanks for adding the description. I have one question after reading the description for the horizontal advection: What do you do in regions of diverging or converging latteral movement? Is the elevation interpolated from the last time step, and then uplift/subsidence is addded if there is any? In this sense the vertical and horizontal movement of the surface is applied *kinematically* (if this is the right word?) at each time step, and then erosion and deposition occur leading to an update in the surface elevation?
