### 2020-12-01 eca7195
Update tecto.rst
