### 2020-12-01 Geodels/gospl#94 @tristan-salles
That's right, I have added it in the tectonic forcing description in the technical guide.
