### 2020-12-02 #2804 @tristan-salles
Thanks @johnjarmitage. Following your comments I have updated the technical guidelines. If you are happy with the changes I will close the open issue on GitHub (@kbarnhart).
