### 2020-12-02 #2804 @johnjarmitage
@tristan-salles I closed the issue, as I thought the technical note was good enough. While I am here, I wonder if the JOSS article should mention the divergence between gospl, pyBadlands (or just Badlands) and eSCAPE. I think all of these codes use a similar approach, therefore why do we need gospl? What differentiates it from the others? Is it a simplified Badlands (no carbonates etc)?  Does eSCAPE only do global simulations?
