+++ paper/paper.md
@@ -38,6 +38,8 @@ The emergence of forward modelling of landscapes and sedimentary systems has pro
 Yet, we are still missing a tool to evaluate global scale patterns of paleo-Earth surface and the interactions with both the atmosphere, the hydrosphere, the tectonics and mantle dynamics.  **gospl** is designed to address parts of this gap and to make the link between deep Earth, climate and Earth' surface. It pioneers innovative and efficient techniques to estimate and predict ancient landscapes and sedimentary successions at global scale and over deep time [@Salles:18b]. It can be used to test different hypothesis related to Earth past evolution and to characterise the role of several drivers such as precipitation, dynamic topography, sea-level on Earth landscape evolution and sedimentary basins formation.
 
 
+Applications wise, the major difference between **gospl** and existing landscape evolution models (*e.g.* Badlands [@Salles:18], LandLab [@Hobley:17] or FastScape [@Braun:13]) is that it focusses on global scale and long term (several millions of years) problems and as such offers new possibilities to understand deep time, all-Earth surface evolution and associated stratigraphy formation. It is designed to be used as an independent tool but can also inform boundary conditions often required in aforementioned regional models (like incoming upstream sediment fluxes when considering continental scale catchments). In terms of implementation, the flow routing technique and river incision are based on a parallel approach and use a multiple flow direction algorithm both of which are not possible within Badlands. It greatly improves runtime performance and allows to better represent rivers routing in flat regions. Compared to eSCAPE [Salles:18b] which has the capability to be run at both regional and global scales, it allows for (1) horizontal advection to be accounted for and (2) multiple lithologies and associated compaction to be represented. While eSCAPE can be used at both scales, it fails short when one wants to evaluate the impacts of long term plate motions on landscape evolution and associated sedimentation. Therefore **gospl** is better suited for global scale applications requiring to account for both vertical and horizontal displacements. In addition to its ability to consider different sediment types, **gospl** can also be used with different mesh resolutions enabling better representation of surface processes in  regions of interest (*e.g.* specific basins, continental regions) while using lower resolutions to save memory allocation in other parts (deep marine regions for example).    
+
 # Package summary
 
 
