### 2020-12-03 Geodels/gospl#104 @cmshobe
JOSS review: very minor points
This issue is part of my JOSS review: openjournals/joss-reviews#2804.

Hi @tristan-salles, congratulations on a really nice piece of software! I have used it through Docker and run the example notebooks without trouble (except for occasionally running out of memory on my machine, but that's my problem not yours).

I also find the technical guide very thorough after the changes suggested by John. I have just a couple of small reference additions to request for the paper (they may also be relevant to the technical guide). These are very minor:

- Rivenaes (1992, Basin Research) should be cited for the two-fraction diffusion model.
- Sclater and Christie (1980, JGR) should be cited for the compaction in addition to Yuan et al., since the latter is really presenting a clever shortcut to approximating the compaction law reported by the former.
- Might you consider citing one of the foundational stream power papers given that you provide citations for the other process components? Howard et al (1994, JGR) would be one good option.

I wish I had more to offer in this review, but the software is working well and the technical guide is very thorough. `gospl` will be a big asset to the community.
