+++ paper/paper.md
@@ -37,7 +37,6 @@ The emergence of forward modelling of landscapes and sedimentary systems has pro
 
 Yet, we are still missing a tool to evaluate global scale patterns of paleo-Earth surface and the interactions with both the atmosphere, the hydrosphere, the tectonics and mantle dynamics.  **gospl** is designed to address parts of this gap and to make the link between deep Earth, climate and Earth' surface. It pioneers innovative and efficient techniques to estimate and predict ancient landscapes and sedimentary successions at global scale and over deep time [@Salles:18b]. It can be used to test different hypothesis related to Earth past evolution and to characterise the role of several drivers such as precipitation, dynamic topography, sea-level on Earth landscape evolution and sedimentary basins formation.
 
-
 Applications wise, the major difference between **gospl** and existing landscape evolution models (*e.g.* Badlands [@Salles:18], LandLab [@Hobley:17] or FastScape [@Braun:13]) is that it focusses on global scale and long term (several millions of years) problems and as such offers new possibilities to understand deep time, all-Earth surface evolution and associated stratigraphy formation. It is designed to be used as an independent tool but can also inform boundary conditions often required in aforementioned regional models (like incoming upstream sediment fluxes when considering continental scale catchments). In terms of implementation, the flow routing technique and river incision are based on a parallel approach and use a multiple flow direction algorithm both of which are not possible within Badlands. It greatly improves runtime performance and allows to better represent rivers routing in flat regions. Compared to eSCAPE [Salles:18b] which has the capability to be ran at both regional and global scales, it allows for (1) horizontal advection to be accounted for and (2) multiple lithologies and associated compaction to be represented. While eSCAPE can be used at both scales, it fails short when one wants to evaluate the impacts of long term plate motions on landscape evolution and associated sedimentation. Therefore **gospl** is better suited for global scale applications requiring to account for both vertical and horizontal displacements. In addition to its ability to consider different sediment types, **gospl** can also be used with different mesh resolutions enabling better representation of surface processes in  regions of interest (*e.g.* specific basins, continental regions) while using lower resolutions to save memory allocation in other parts (deep marine regions for example).    
 
 # Package summary
@@ -46,11 +45,11 @@ Applications wise, the major difference between **gospl** and existing landscape
 **gospl** is able to simulate global-scale forward models of landscape evolution, dual-lithology (coarse and fine) sediment routing and stratigraphic history forced with deforming plate tectonics, paleotopographies and paleoclimate reconstructions. It relates the complexity of the triggers and responses of sedimentary processes from the complete sediment routing perspective accounting for different scenarii of plate motion, tectonic uplift/subsidence, climate, geodynamic and sedimentary conditions.
 
 The following physical processes are considered:
-- river incision using the stream power law based on a multiple flow direction approach and an implicit parallel implicit drainage area method [@Richardson:14],
+- river incision using the stream power law [@Howard:94] based on a multiple flow direction approach and an implicit parallel implicit drainage area method [@Richardson:14],
 - inland river deposition in depressions computed using priority-flood techniques [@Barnes:14],
-- dual-lithology marine deposition at river mouth based on a diffusion algorithm,
+- dual-lithology marine deposition at river mouth based on a diffusion algorithm [@Rivenaes:92],
 - hillslope processes in both marine and inland areas [@Salles:19], and
-- sediment compaction as stratigraphic layers geometry and properties changes [@Yuan:19].
+- sediment compaction as stratigraphic layers geometry and properties changes [@Sclater:80; @Yuan:19].
 
 As mentioned previously, **gospl** can be forced with spatially and temporally varying tectonics (horizontal and vertical displacements) and climatic forces (temporal and spatial precipitation changes and sea-level fluctuations).
 
