+++ paper/paper.bib
@@ -8,6 +8,39 @@
   doi = {10.1016/j.sedgeo.2018.06.007}
 }
 
+@article{Sclater:80,
+author = {Sclater, John G. and Christie, P. A. F.},
+title = {{Continental stretching: An explanation of the Post-Mid-Cretaceous subsidence of the central North Sea Basin}},
+journal = {Journal of Geophysical Research: Solid Earth},
+volume = {85},
+number = {B7},
+pages = {3711-3739},
+doi = {https://doi.org/10.1029/JB085iB07p03711},
+year = {1980}
+}
+
+@article{Howard:94,
+author = {Howard, Alan D. and Dietrich, William E. and Seidl, Michele A.},
+title = {Modeling fluvial erosion on regional to continental scales},
+journal = {Journal of Geophysical Research: Solid Earth},
+volume = {99},
+number = {B7},
+pages = {13971-13986},
+doi = {10.1029/94JB00744},
+year = {1994}
+}
+
+@article{Rivenaes:92,
+author = {Rivenaes, Jan C.},
+title = {Application of a dual-lithology, depth-dependent diffusion equation in stratigraphic simulation},
+journal = {Basin Research},
+volume = {4},
+number = {2},
+pages = {133--146},
+doi = {10.1111/j.1365-2117.1992.tb00136.x},
+year = {1992}
+}
+
 @article{Helland:16,
     author = {Helland-Hansen, W. and Somme, T. O. and Martinsen, O. J. and Lunt, I. and Thurmond, J.},
     title = "{D}eciphering {E}arth's {N}atural {H}ourglasses: {P}erspectives {O}n {S}ource-{T}o-{S}ink {A}nalysis",
