### 2020-12-03 7c31d19
add references suggested by JOSS reviewer in the doc
