### 2020-12-03 Geodels/gospl#104 @tristan-salles
Hi @cmshobe, thank you for agreeing to do the review, really appreciate it! I have made the modifications as suggested.
