### 2020-12-06 #2804 @kbarnhart
@johnjarmitage and @cmshobe - thanks for your reviews.
@tristan-salles - thanks for addressing reviewer issues.

Based on the discussion, it seems like there are no more reviewer comments. If this is the case, could both reviewers:
(a) look through their checklists (earlier on this thread) and check off completed boxes. If there are any remaining comments related to checking off an item, of course mention it so it can be addressed.
(b) If all check boxes are checked, indicate that you are happy with the current state of the submission.

Thanks!
