### 2020-12-09 #2804 @cmshobe
I am sorry for being late-- This is my first JOSS review and I had forgotten about the formal checklist! I am ready to see this published. This is a fantastic software package and I look forward to seeing where people go with it in the future.
