### 2020-12-09 #2804 @kbarnhart
@johnjarmitage and @cmshobe thanks for your affirmation and your reviews. (no worries @cmshobe)

@tristan-salles at this point I'll do final copyediting checks and let you know the next steps. It may take a few days because of AGU.
