+++ requirements.txt
@@ -8,7 +8,7 @@ pandas>=1.1.2
 ruamel.yaml==0.16.12
 fastfunc==0.2.3
 meshio>=4.2.2
-meshplex==0.13.4
+meshplex==0.15.5
 pre-commit>=2.7.1
 vtk>=8.1.2
 numpy-indexed==0.3.5
