+++ paper/paper.bib
@@ -15,7 +15,7 @@ journal = {Journal of Geophysical Research: Solid Earth},
 volume = {85},
 number = {B7},
 pages = {3711-3739},
-doi = {https://doi.org/10.1029/JB085iB07p03711},
+doi = {10.1029/JB085iB07p03711},
 year = {1980}
 }
 
