+++ paper/paper.bib
@@ -108,7 +108,7 @@ year = {1992}
 }
 
 @Inbook{Tucker:01,
-  author = {Tucker, Gregory and Lancaster, Stephen  and Gasparini, Nicole and Bras, Rafael},
+  author = {Tucker, G E and Lancaster, Stephen  and Gasparini, Nicole and Bras, Rafael},
   editor = {Harmon, Russell S.  and Doe, William W.},
   title = "{T}he {C}hannel-{H}illslope {I}ntegrated {L}andscape {D}evelopment {M}odel ({CHILD})",
   bookTitle = {Landscape Erosion and Evolution Modeling},
@@ -131,7 +131,7 @@ year = {1992}
 }
 
 @article{Braun:13,
-  author = {Braun, Jean and  Willett, Sean D.},
+  author = {Braun, J and  Willett, Sean D.},
   title = "A very efficient {O}(n), implicit and parallel method to solve the stream power equation governing fluvial incision and landscape evolution",
   journal = {Geomorphology},
   volume = {180--181},
@@ -142,7 +142,7 @@ year = {1992}
 }
 
 @article{Tucker:10,
-  author = {Tucker, Gregory E. and Hancock, Gregory R.},
+  author = {Tucker, G. E. and Hancock, Gregory R.},
   title = {Modelling landscape evolution},
   journal = {Earth Surface Processes and Landforms},
   volume = {35},
@@ -218,7 +218,7 @@ year = {1992}
 }
 
 @article{Hobley:11,
-  author = {Hobley, Daniel E. J. and Sinclair, Hugh D. and Mudd, Simon M. and Cowie, Patience A.},
+  author = {Hobley, D. E. J. and Sinclair, Hugh D. and Mudd, Simon M. and Cowie, Patience A.},
   title = {Field calibration of sediment flux dependent river incision},
   journal = {Journal of Geophysical Research: Earth Surface},
   volume = {116},
