+++ paper/paper.bib
@@ -130,6 +130,17 @@ year = {1992}
   doi = {10.5194/esurf-5-21-2017}
 }
 
+@Article{barnhart2020short,
+AUTHOR = {Barnhart, K. R. and Hutton, E. W. H. and Tucker, G. E. and Gasparini, N. M. and Istanbulluoglu, E. and Hobley, D. E. J. and Lyons, N. J. and Mouchene, M. and Nudurupati, S. S. and Adams, J. M. and Bandaragoda, C.},
+TITLE = {Short communication: Landlab v2.0: a software package for Earth surface dynamics},
+JOURNAL = {Earth Surface Dynamics},
+VOLUME = {8},
+YEAR = {2020},
+NUMBER = {2},
+PAGES = {379--397},
+DOI = {10.5194/esurf-8-379-2020}
+}
+
 @article{Braun:13,
   author = {Braun, J and  Willett, Sean D.},
   title = "A very efficient {O}(n), implicit and parallel method to solve the stream power equation governing fluvial incision and landscape evolution",
