### 2020-12-12 dc88492
copyedits
commas, lists, reference formatting, add second landlab reference.
