+++ paper/paper.bib
@@ -108,7 +108,7 @@ year = {1992}
 }
 
 @Inbook{Tucker:01,
-  author = {Tucker, G E and Lancaster, Stephen  and Gasparini, Nicole and Bras, Rafael},
+  author = {Tucker, G. E. and Lancaster, Stephen  and Gasparini, Nicole and Bras, Rafael},
   editor = {Harmon, Russell S.  and Doe, William W.},
   title = "{T}he {C}hannel-{H}illslope {I}ntegrated {L}andscape {D}evelopment {M}odel ({CHILD})",
   bookTitle = {Landscape Erosion and Evolution Modeling},
@@ -142,7 +142,8 @@ DOI = {10.5194/esurf-8-379-2020}
 }
 
 @article{Braun:13,
-  author = {Braun, J and  Willett, Sean D.},
+  author = {Braun, J.
+   and  Willett, Sean D.},
   title = "A very efficient {O}(n), implicit and parallel method to solve the stream power equation governing fluvial incision and landscape evolution",
   journal = {Geomorphology},
   volume = {180--181},
