### 2020-12-12 #2804 @kbarnhart
@tristan-salles here are the last steps:
- [ ] Merge PR I've opened with copyediting. Full disclosure, I've added a reference that is my own work, but this is because Landlab now has two core references.

Once that is done, please run `@whedon check references` if no invalid DOIs come up, please do the following:

I will then move forward with recommending the submission be accepted and published
