+++ setup.py
@@ -38,7 +38,7 @@ except ImportError:
     pass
 
 # in development set version to none and ...
-PYPI_VERSION = "0.2.0"
+PYPI_VERSION = "0.2.1"
 
 # Place install_requires into the text file "requirements.txt"
 with open("requirements.txt") as f2:
