+++ docs/conf.py
@@ -44,9 +44,9 @@ author = "Tristan Salles"
 # release = version
 
 # The short X.Y version
-version = "0.2.0"
+version = "0.2.1"
 # The full version, including alpha/beta/rc tags
-release = "0.2.0"
+release = "0.2.1"
 
 
 # -- General configuration ---------------------------------------------------
