+++ conda-env.yml
@@ -21,5 +21,5 @@ dependencies:
   - compilers
   - pip:
     - fastfunc
-    - meshplex==0.13.4
+    - meshplex
     - gospl
