### 2020-12-13 #2804 @tristan-salles
Hi @kbarnhart,

Thanks for the suggestions!

I have merged PR and ran @whedon check references. I made a tagged release of the software, with the following Zenodo reference:
- Tristan Salles, Claire Mallard, & Sabin Zahirovic. (2020, December 13). gospl: Global Scalable Paleo Landscape Evolution (Version 0.2.1). Zenodo. http://doi.org/10.5281/zenodo.4319332

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4319332.svg)](https://doi.org/10.5281/zenodo.4319332)
