### 2020-12-13 be67870
update documentation paper DOI
