+++ docs/refs.bib
@@ -1,0 +1,283 @@
+@article{Nyberg:18,
+  author = { Nyberg, B. and  Helland-Hansen, W. and Gawthorpe, R. L.  and  Sandbakken, P. and  Haug Eide, C. and  Somme, T. and  Hadler-Jacobsen, F. and  Leiknes, S.},
+  title = {Revisiting morphological relationships of modern source-to-sink segments as a first-order approach to scale ancient sedimentary systems},
+  journal = {Sedimentary Geology},
+  volume = {373},
+  pages = {111--133},
+  year = {2018},
+  doi = {10.1016/j.sedgeo.2018.06.007}
+}
+
+@article{Sclater:80,
+author = {Sclater, John G. and Christie, P. A. F.},
+title = {{Continental stretching: An explanation of the Post-Mid-Cretaceous subsidence of the central North Sea Basin}},
+journal = {Journal of Geophysical Research: Solid Earth},
+volume = {85},
+number = {B7},
+pages = {3711-3739},
+doi = {10.1029/JB085iB07p03711},
+year = {1980}
+}
+
+@article{Howard:94,
+author = {Howard, Alan D. and Dietrich, William E. and Seidl, Michele A.},
+title = {Modeling fluvial erosion on regional to continental scales},
+journal = {Journal of Geophysical Research: Solid Earth},
+volume = {99},
+number = {B7},
+pages = {13971-13986},
+doi = {10.1029/94JB00744},
+year = {1994}
+}
+
+@article{Rivenaes:92,
+author = {Rivenaes, Jan C.},
+title = {Application of a dual-lithology, depth-dependent diffusion equation in stratigraphic simulation},
+journal = {Basin Research},
+volume = {4},
+number = {2},
+pages = {133--146},
+doi = {10.1111/j.1365-2117.1992.tb00136.x},
+year = {1992}
+}
+
+@article{Helland:16,
+    author = {Helland-Hansen, W. and Somme, T. O. and Martinsen, O. J. and Lunt, I. and Thurmond, J.},
+    title = "{D}eciphering {E}arth's {N}atural {H}ourglasses: {P}erspectives {O}n {S}ource-{T}o-{S}ink {A}nalysis",
+    journal = {Journal of Sedimentary Research},
+    volume = {86},
+    number = {9},
+    pages = {1008--1033},
+    year = {2016},
+    doi = {10.2110/jsr.2016.56}
+}
+
+@article{Richardson:14,
+  author = {Richardson, A. and Hill, C. N. and Perron, J. T.},
+  title = "{IDA}: {A}n implicit, parallelizable method for calculating drainage area",
+  journal = {Water Resources Research},
+  volume = {50},
+  number = {5},
+  pages = {4110-4130},
+  doi = {10.1002/2013WR014326},
+  year = {2014}
+}
+
+@article{Barnes:14,
+   author = {Barnes, R. and Lehman, C. and Mulla, D.},
+    title = "Priority-flood: {A}n optimal depression-filling and watershed-labeling algorithm for digital elevation models",
+  journal = {Computers \& Geosciences},
+     year = {2014},
+   volume = {62},
+    pages = {117-127},
+      doi = {10.1016/j.cageo.2013.04.024}
+}
+
+@article{Yuan:19,
+  author = {Yuan, X.P. and Braun, J. and Guerit, L. and Simon, B. and Bovy, B. and Rouby, D. and Robin, C. and Jiao, R.},
+  title = "Linking continental erosion to marine sediment transport and deposition: {A} new implicit and {O(N)} method for inverse analysis",
+  journal = {Earth and Planetary Science Letters},
[…]
