### 2020-12-13 #2804 @kbarnhart
@tristan-salles this is now in the hands of the JOSS handling editor in chief who will address final publication steps.

Thanks @johnjarmitage and @cmshobe for your reviews.
