### 2020-12-14 #2804 @kthyng
Hi @tristan-salles! I'll be taking over to get your publication finalized. Looks like everything is in place. I have one comment from the paper. They are your own publications, but I see one called "eSCAPE" and one called "ESCAPE" but apparently should both be "eSCAPE". That one extra capital letter gives a really different feeling to the word. Would you like to update this in your references?
