+++ paper/paper.bib
@@ -86,7 +86,7 @@ year = {1992}
 
 @article{Salles:19,
   author = {Salles, T.},
-  title = "e{SCAPE}: {R}egional to {G}lobal {S}cale {L}andscape {E}volution {M}odel v2.0",
+  title = "{eSCAPE}: {R}egional to {G}lobal {S}cale {L}andscape {E}volution {M}odel v2.0",
   journal = {Geoscientific Model Development},
   volume = {12},
   year = {2019},
